tkinter/requirements.txt
pylint
black
Pillow